Bump some dependencies
Started off wanting pretty-simple-4 in order to indent a whole expression. Required newer prettyprinter, which required newer dhall, which added an instance for Int, which clashed with our instance, which used Natural instead of Integer. So this prompts us to correctly indicate which values should be signed/unsigned. Seems like a good thing in the long run.