Delay handling of `pointerleave` events
We had an issue whereby the dict lookup in the leave handler could fail because `pointerleave` followed `pointerdown` very closely, and thus the latter overtook our `PointerUp` Elm event. By making the `pointerleave` handling also jump through an Elm event, we restore the correct order.

One symptom of this was that if a button was pressed and released very quickly, it would stay pressed.

This is a shortcoming of the Elm architecture really (that we can't modify the model directly in event handlers, only send messages).