Add freeze file
Seeing as this is all a bit brittle with the dependency bound overrides etc. we ought to lock everything down.

We manually remove `monpad -release` from the freeze file so we can continue to build in both modes. An alternative might be to keep the untampered file, but rename and point to it explicitly from the build script.