Revert 8169c72: "Use a Map rather than List for the layout elements"
Obviously, in hindsight, that change broke any layouts where we rely on order in the input list to draw some elements on top of others.

We do fix the inconsistency mentioned in that commit message (updates affected all matching elements in the frontend, but only the first in the backend). So it is only performance that we lose.