Simplify `onPong`
Parallel composition in stateful monads is not supported by Streamly.

Reverts 6ba3937, and therefore unfortunately 0f54b13, but we'll open an issue for that.

This possibly explains why the ping indicator was brkoen on Windows - it wasn't guaranteed to work anywhere. But we haven't tested this version on Windows yet.