Validate usernames once, then use a waiting set for valid WS usernames
Old version is broken because we try to modify the set of users twice. The second time, when opening the websocket, we fail because it's already in the set. So we change our approach to avoid any awkward concurrency issues.

We go beyond the original purpose of this PR in that we also defend against differences between the WS and HTTP username params. e.g. if a user modifies the served JS.