Use `cabal list-bin` to make locating build products more robust
To be fair, I'm pretty sure the command didn't exist when I wrote the first iteration of this build script.

For one thing, this change ensures that the build script doesn't fail when we've been building with multiple GHC versions.

It's unclear to me exactly when, if ever, `list-bin` would return multiple results. But the documentation seems to suggest that it's possible. So we handle it.