Remove the "monpad" prefix from all routes
This means that e.g. one now connects to `localhost:8000 ` instead of `localhost:8000/monpad`.

Breaking changes:
- Obviously, any connection must omit the "monpad" prefix.
- In `--ext-ws` mode, one needs to add "/ws" to the route.
- All references to locally served files (images used in layouts, audio files used in server updates, etc.) need to be modified to remove the prefix.

Note that we had to swap ordering of HTTP and WS APIs in `server`, because `Raw` acts as a catch-all.

Note also that we could have opted to keep the prefix just for assets, thus mostly retaining backwards compatibility, but it would have been a slightly awkward vestige, and if we're ever going to remove it, we may as well do it now, so that we can make all downstream changes at once.