Add `dump-html` option
Note that we now specify that we use UTF8 encoding in the meta tag. This wasn't previously a concern since Warp seemingly sets it in `Content-Type` in all response headers (which seems reasonable, as it's apparently [the only option](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta)). But other servers may not do so.

We _could_ potentially now drop `ext-ws` mode as obsolete, since this new mode provides greater flexibility. But it _might_ still be a useful middle ground to someone, and it doesn't cost much to maintain for now.