Use Google Closure Compiler if available
Results are pretty impressive: on `main.js`, the difference is less than 10%, but on the much larger `elm.js`, we go from 306KB un-optimised, and 230K with the Haskell minifier, to a mere 76K with Closure. This is with version `v20230228`, installed from the Arch User Repository.

Note that this is without the "advanced optimisations" being switched on. They don't work out of the box and would require some manual annotations, but this could be worth looking in to. Indeed, the maintainer of Miso, which is very similar to Elm, [claims to have gotten this working](https://news.ycombinator.com/item?id=14693698).

We had to make some light modifications due to the fact that it (and any other command-line minifiers I've seen) doesn't seem to like modifying in-place.