Update to `fsnotify-0.4` and use new-style Streamly streams
It's hard to avoid doing this in lockstep. The new `fsnotify` has dropped its built-in debouncing support. We have simpler and more robust support via `groupByTime`, but this only supports the new Streamly API. On the other hand, using the new API requires that we use at least `streamly-fsnotify-2.0`, which doesn't support `fsnotify-0.3`.