Add an option to counteract use of a deadzone (e.g. by Steam)
Note that internally we allow the deadzone to be a vector. This is more flexible and actually very slightly easier to implement. But the CLI only allows a scalar, i.e. using the same value for the x and y coordinates.

We also could have made it non-Maybe and defaulted to 0. But at some point we might want for example to display something if it is in use, and checking a floating point number against 0 is best avoided in general. Besides we'd want to distinguish between it being unset and a user having set it manually to 0.