Avoid unnecessary optimised Elm rebuilds
Sometimes (unclear why not always) we'd get stuck in a loop where the Elm rule would always re-run even though there were no changes. This was because the rule calling `elm make` would depend on a temporary file which it had itself created.

This isn't the cleanest solution, but it should work.