Support building with GHC 9.12
Note that we can't add the SRP in the build script due to https://github.com/haskell/cabal/issues/8024#issuecomment-1805454403. Luckily in this instance, it's possible to get things working by fiddling with bounds and constraints, and nothing else is forcing us to use `time-1.14`, which requires actual code changes for Dhall.